requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
cloudpickle==0.6.1
cycler==0.10.0
dask==0.20.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
graphviz==0.10.1
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
more-itertools==4.3.0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.7.0
Pygments==2.2.0
pyparsing==2.3.0
pytest==3.9.3
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.4.2
scikit-image==0.14.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
